add repository to POM file
The "Pentaho Aggregate Designer Algorithm" package (which is a child dependency of the Hive libraries) is not hosted on Maven.  Instead, it's hosted on conjars.org.  Previously, Maven knew where to pull that library.  But that site seems to be experiencing some disruption, and Maven's pointer is out of date.  I added a repo location that will hopefully point Maven to the right place.  (This situation feels a bit unstable so we may have to change again in the near future.)